+++ lib/src/forwarding.dart
@@ -2,18 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library file.src.forwarding;
-
-import 'dart:async';
-import 'dart:convert';
-
-import 'package:file/src/io.dart' as io;
-import 'package:file/file.dart';
-import 'package:meta/meta.dart';
-import 'package:path/path.dart' as p;
-
-part 'forwarding/forwarding_directory.dart';
-part 'forwarding/forwarding_file.dart';
-part 'forwarding/forwarding_file_system.dart';
-part 'forwarding/forwarding_file_system_entity.dart';
-part 'forwarding/forwarding_link.dart';
+export 'forwarding/forwarding_directory.dart';
+export 'forwarding/forwarding_file.dart';
+export 'forwarding/forwarding_file_system.dart';
+export 'forwarding/forwarding_file_system_entity.dart';
+export 'forwarding/forwarding_link.dart';